requirements.txt
python-dotenv==1.0.1
pandas==2.2.2
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
openpyxl==3.1.5
loguru==0.7.2
tenacity==9.0.0
PyYAML==6.0.2
